give reload its own file, it's growing
